requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcrypt==4.3.0
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==44.0.2
Django==5.1.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
paramiko==3.5.1
pluggy==1.5.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyspnego==0.11.2
pytest==8.3.4
python-dateutil==2.9.0.post0
pywinrm==0.5.0
redis==5.2.1
requests==2.32.3
requests_ntlm==1.3.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
xmltodict==0.14.2
